+++ service/src/main/java/com/guru/service/parser/impl/ParserAA.java
@@ -1,9 +1,12 @@
 package com.guru.service.parser.impl;
 
+import akka.actor.ActorRef;
+import akka.actor.Props;
 import akka.actor.UntypedActor;
 import akka.event.Logging;
 import akka.event.LoggingAdapter;
 import com.guru.service.RequestData;
+import com.guru.service.actor.processingresult.ProcessingResultOfParserActor;
 import com.guru.service.parser.interf.ParserActor;
 import org.apache.http.impl.client.DefaultHttpClient;
 import parser.aa.AAParser;
@@ -11,6 +14,7 @@ import parser.utils.Account;
 import parser.utils.AccountUtils;
 
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.List;
 
 public class ParserAA extends UntypedActor implements ParserActor {
@@ -28,8 +32,9 @@ public class ParserAA extends UntypedActor implements ParserActor {
             DefaultHttpClient loggedInClient = aaParser.logIn(account.getLogin(), account.getPin(), account.getPassword(), account);
            /*List flights = aaParser.getAmericanAirlines(loggedInClient, account, requestData.getOrigin(), requestData.getDestination(),
                     sdf.parse("04/25/2016"), 1);*/
-            List flights = aaParser.getAmericanAirlines(loggedInClient, account, "SDQ", "NYC", sdf.parse("04/25/2016"), 1);
-            getSender().tell(flights, getSelf());
+            List flights = aaParser.getAmericanAirlines(loggedInClient, account, "SDQ", "NYC", sdf.parse("04/26/2016"), 1);
+            ActorRef processingResultOfParserActor = context().system().actorOf(Props.create(ProcessingResultOfParserActor.class));
+            processingResultOfParserActor.tell(flights, self());
         } else unhandled(message);
     }
