+++ service/src/main/java/com/guru/service/parser/impl/ParserEY.java
@@ -1,9 +1,12 @@
 package com.guru.service.parser.impl;
 
+import akka.actor.ActorRef;
+import akka.actor.Props;
 import akka.actor.UntypedActor;
 import akka.event.Logging;
 import akka.event.LoggingAdapter;
 import com.guru.service.RequestData;
+import com.guru.service.actor.processingresult.ProcessingResultOfParserActor;
 import com.guru.service.parser.interf.ParserActor;
 import org.apache.http.impl.client.DefaultHttpClient;
 import parser.aa.AAParser;
@@ -31,7 +34,8 @@ public class ParserEY extends UntypedActor implements ParserActor {
             List flights = eyParser.getEtihad("04/25/2016", "LAX", "AUH", 1, "E");
            /* List flights = eyParser.getEtihad("04/25/2016", requestData.getOrigin(),
                     requestData.getDestination(), 1, "E");*/
-            getSender().tell(flights, getSelf());
+            ActorRef processingResultOfParserActor = context().system().actorOf(Props.create(ProcessingResultOfParserActor.class));
+            processingResultOfParserActor.tell(flights, self());
         } else unhandled(message);
 
     }
